Allow for optional command execution on backups
This was originally inteded for use of compression, however can be used for pretty much everything, from shipping the files to a remote location to just rm-ing the files for whatever reason.
The only issue right now is that it WILL perform a permission denied error on standard configuration settings if the file did not get a chmod, however since the command might never be performed it is quite likely that the exception will never be encountered for the general population.

Signed-off-by: Geolykt <admin@geolykt.de>